+++ src/cli/swpm/commands/interactive.ts
@@ -34,6 +34,8 @@ const interactive: CommandModule<Record<string, unknown>, OptionsProps> = {
       }),
 
   handler: (yargs) => {
+    if (!cmdr?.cmd) return
+
     if ('latest' in yargs) {
       translateArgs({ yargs, cmdr, flag: '--latest', alias: '-L' })
     }