+++ src/cli/swpm/commands/upgrade.ts
@@ -44,6 +44,8 @@ const upgrade: CommandModule<Record<string, unknown>, OptionsProps> = {
       }),
 
   handler: (yargs) => {
+    if (!cmdr?.cmd) return
+
     if ('save-exact' in yargs) {
       translateArgs({ yargs, cmdr, flag: '--save-exact', alias: '-E' })
     }