+++ src/helpers/debug.ts
@@ -1,6 +1,8 @@
 import { inspect } from 'node:util'
+import { ArgumentsCamelCase } from 'yargs'
+import { CommanderPackage } from '../translator/commander.types'
 
-export const debug = (object: Object) => {
+export const debug = (object: CommanderPackage | ArgumentsCamelCase) => {
   console.log(
     inspect(
       object,