+++ src/helpers/icons.test.ts
@@ -4,7 +4,7 @@ import { getOriginIcon } from './icons.js'
 describe('getOriginIcon()', () => {
   it('should return an empty string if name was not send', () => {
     const expectedResult = ''
-    // @ts-expect-error
+    // @ts-expect-error expected result is intentionally empty
     const result = getOriginIcon()
     expect(result).toBe(expectedResult)
   })
@@ -12,7 +12,7 @@ describe('getOriginIcon()', () => {
   it('should return an empty string if name was not found', () => {
     const expectedResult = ''
     const name = 'not'
-    // @ts-expect-error
+    // @ts-expect-error name is intentionally a non existent icon
     const result = getOriginIcon(name)
     expect(result).toBe(expectedResult)
   })